pyproject.toml
[project]
name = "kafkaproject"
version = "0.1.0"
description = ""
authors = [
    {name = "DanielArama0",email = "DanielArama95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python (>=2.1.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "kafkaproject", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
